+++ src/common.module.scss
@@ -1,3 +1,6 @@
+$tBlue: #1d9bf0;
+$tGray: #868a8c;
+
 .input {
   border: solid 1px #cfd9de;
   border-radius: 5px;
