+++ src/components/Header/index.module.scss
@@ -1,7 +1,6 @@
 .header {
   height: 50px;
   text-align: center;
-  background-color: #ccc;
 }
 
 .closeIcon {
