+++ src/containers/Register/index.js
@@ -18,6 +18,7 @@ const ACCOUNT_TYPE = {
  *  register page
  */
 const Regsister = () => {
+  const [form] = Form.useForm();
   const [formData] = useState({
     name: '',
     tel: '',
@@ -35,22 +36,47 @@ const Regsister = () => {
     setAccountType(ACCOUNT_TYPE.TEL);
   };
 
+  const onClickNextStep = async () => {
+    const validata = await form.validateFields();
+    if (validata) {
+      console.log(validata);
+    }
+  };
+
   return (
     <div>
       <Header />
       <div className={style.form}>
         <div className={style.formTitle}>Create your account</div>
-        <Form initialValues={formData} className={style.formContainer}>
-          <Form.Item name="name">
+        <Form form={form} initialValues={formData} className={style.formContainer}>
+          <Form.Item
+            name="name"
+            rules={[{
+              required: true,
+              message: 'Please enter your name.',
+            }]}
+          >
             <Input placeholder="Name" className={style.input} />
           </Form.Item>
           {accountType === ACCOUNT_TYPE.TEL && (
-          <Form.Item name="tel">
+          <Form.Item
+            name="tel"
+            rules={[{
+              required: true,
+              message: 'Please enter your phone number.',
+            }]}
+          >
             <Input placeholder="Phone" className={style.input} />
           </Form.Item>
           )}
           {accountType === ACCOUNT_TYPE.EMAIL && (
-          <Form.Item name="email">
+          <Form.Item
+            name="email"
+            rules={[{
+              required: true,
+              message: 'Please enter your email.',
+            }]}
+          >
             <Input placeholder="Email" className={style.input} />
           </Form.Item>
           )}
@@ -65,7 +91,7 @@ const Regsister = () => {
         </Form>
       </div>
       <div className={style.footer}>
-        <Button className={style.footerButton}>Next</Button>
+        <Button className={style.footerButton} onClick={onClickNextStep}>Next</Button>
       </div>
     </div>
   );
